requirements.txt
# Core Flask extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Limiter==3.5.0

# Security & Authentication
Werkzeug==3.0.1
python-dotenv==1.0.0
email-validator==2.1.0

# Production Server
gunicorn==21.2.0
gevent==23.9.1

# Database (PostgreSQL)
psycopg2-binary==2.9.9

# File handling & Validation
Pillow==10.1.0
python-magic==0.4.27

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Monitoring & Logging
sentry-sdk[flask]==1.38.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Code Quality
flake8==7.0.0
black==23.11.0
